Enhance action discovery with consolidated search and keyword support
Consolidates search logic from ActionPopover and ActionSelectorNode into
a shared useActionSearch composable and ActionZone component.

Key improvements:
- Direct Operation Discovery: Users can find operations (e.g., 'Clear')
  without knowing the parent Action Type (e.g., 'Assignment').
- Keyword Support: Added a keywords registry for aliases like 'email'
  pointing to 'Notify'.
- Unified Ranking: Weighted scoring considering labels, descriptions,
  and keywords.
- Logic Consolidation: Removed significant duplication between components.
- Backward Compatibility: Preserves existing scoped search and node
  creation payloads.

Includes a detailed audit report in audit_report.md.
