+++ flexirule/public/js/flexirule/rule_builder/components/nodes/ActionSelectorNode.vue
@@ -1,16 +1,13 @@
 <script setup>
-import { ref, computed, onMounted } from "vue";
+import { ref, computed } from "vue";
 import { Handle, Position } from "@vue-flow/core";
 import {
-	ACTION_TYPE_CONTRACT,
-	PROCESS_REGISTRY,
-	getActionTypeOptions,
 	getOperationOptions,
-	loadContractsFromBackend,
 	isTerminalAction,
 } from "../../../core/contracts";
 import { useStore } from "../../stores";
 import { mapActionTypeToNodeType } from "../../composables/useActionTypeMapper";
+import ActionZone from "../ActionZone.vue";
 
 const props = defineProps(["data", "label", "id", "selected", "sourcePosition", "targetPosition"]);
 const store = useStore();
@@ -31,206 +28,20 @@ const selectedPreset = ref({
 	selected_label: "Process",
 });
 const customLabel = ref("");
-const searchQuery = ref("");
-const showResults = ref(false);
-const processOperations = ref([]);
-const selectedIndex = ref(-1);
 
-// ... (rest of the script)
-// Fuzzy match helper — matches each query word independently against the text
-// Fuzzy match helper — supports acronyms and word-start matching
-function fuzzyMatch(text, query) {
-	if (!query) return true;
-	const lowerText = text.toLowerCase();
-	const q = query.toLowerCase().trim();
-
-	// 1. Simple substring match
-	if (lowerText.includes(q)) return true;
-
-	// 2. Acronym match (e.g. "dq" -> "Document Query")
-	const words = lowerText.split(/[\s_-]+/).filter(Boolean);
-	const acronym = words.map((w) => w[0]).join("");
-	if (acronym.includes(q)) return true;
-
-	// 3. Multi-word match (e.g. "doc query" -> "Document Query Records")
-	const qWords = q.split(/\s+/).filter(Boolean);
-	return qWords.every((qw) => words.some((w) => w.startsWith(qw) || w.includes(qw)));
+function onActionSelect(item) {
+	selectedPreset.value = {
+		action_type: item.action_type || "Process",
+		operation: item.operation || null,
+		process_name: item.process_name || null,
+		selected_label: item.label || item.value || "Process",
+	};
 }
 
-const actionTypes = computed(() => {
-	return getActionTypeOptions()
-		.filter((t) => t && t !== "Entry Action" && t !== "Start")
-		.map((t) => {
-			const contract = ACTION_TYPE_CONTRACT[t] || {};
-			return {
-				label: __(t),
-				value: t,
-				actionType: t,
-				icon: contract.css?.icon || "fa fa-cog",
-				color: contract.css?.color || "#6b7280",
-				description: contract.description || "",
-				category: __(contract.category || "Other"),
-			};
-		});
-});
-
-// Filtered results for fuzzy search
-const filteredResults = computed(() => {
-	const q = searchQuery.value;
-	const results = [];
-
-	// 1. Filter action types and their native operations
-	actionTypes.value.forEach((t) => {
-		const contract = ACTION_TYPE_CONTRACT[t.value] || {};
-		const ops = getOperationOptions(t.value);
-
-		// Match the action type itself
-		const matchAction =
-			fuzzyMatch(t.label, q) || fuzzyMatch(t.description, q) || fuzzyMatch(t.value, q);
-		if (matchAction) {
-			results.push({ type: "header", label: t.label });
-			results.push({ type: "action", ...t });
-		}
-
-		// Match operations within this action type
-		const matchingOps = ops.filter(
-			(op) => fuzzyMatch(op.label || op.value, q) || fuzzyMatch(op.value, q)
-		);
-		if (matchingOps.length) {
-			if (!matchAction) {
-				results.push({ type: "header", label: t.label });
-			}
-			matchingOps.forEach((op) => {
-				results.push({
-					type: "op",
-					label: `${t.label} → ${__(op.label || op.value)}`,
-					value: t.value,
-					operation: op.value,
-					process_name: t.value === "Process" ? op.process_name || null : null,
-					icon: t.icon,
-					color: t.color,
-					description: t.description,
-				});
-			});
-		}
-	});
-
-	// 2. Filter process operations (backend processes)
-	const matchingProcessOps = processOperations.value.filter(
-		(op) =>
-			fuzzyMatch(op.label, q) ||
-			fuzzyMatch(op.process, q) ||
-			fuzzyMatch(op.description || "", q)
-	);
-
-	if (matchingProcessOps.length) {
-		results.push({ type: "header", label: __("Process Operations") });
-		matchingProcessOps.forEach((op) =>
-			results.push({
-				type: "process_op",
-				label: `${op.process} → ${op.label}`,
-				value: "Process",
-				process_name: op.process,
-				operation: op.operation,
-				icon: "fa fa-cog",
-				color: "#8b5cf6",
-				description: op.description || "",
-			})
-		);
-	}
-
-	return results;
-});
-
-async function loadProcessOperations() {
-	await loadContractsFromBackend();
-	if (Array.isArray(PROCESS_REGISTRY) && PROCESS_REGISTRY.length) {
-		processOperations.value = PROCESS_REGISTRY.flatMap((proc) =>
-			(proc.operations || [])
-				.filter((op) => op.enabled !== 0 && op.visible_in_builder !== 0 && op.func_name)
-				.map((op) => ({
-					process: proc.name,
-					module: proc.module,
-					operation: op.func_name,
-					label: op.label || op.func_name,
-					description: op.description || "",
-				}))
-		);
-		return;
-	}
-	try {
-		const res = await frappe.call({
-			method: "flexirule.ruleflow.api.get_all_process_operations",
-		});
-		processOperations.value = res.message || [];
-	} catch (e) {
-		processOperations.value = [];
-	}
-}
-
-function selectItem(item) {
-	if (item.type === "header") return;
-	if (item.type === "process_op") {
-		selectedPreset.value = {
-			action_type: "Process",
-			operation: item.operation,
-			process_name: item.process_name,
-			selected_label: item.label || item.operation || "Process",
-		};
-	} else if (item.type === "op") {
-		selectedPreset.value = {
-			action_type: item.value || "Process",
-			operation: item.operation || null,
-			process_name: item.process_name || null,
-			selected_label: item.label || item.operation || item.value || "Process",
-		};
-	} else {
-		selectedPreset.value = {
-			action_type: item.value || "Process",
-			operation: null,
-			process_name: null,
-			selected_label: item.label || item.value || "Process",
-		};
-	}
-
-	searchQuery.value = item.label;
-	showResults.value = false;
-}
-
-function onSearchFocus() {
-	showResults.value = true;
-}
-
-function onSearchKeydown(e) {
-	if (!showResults.value || !filteredResults.value.length) return;
-
-	if (e.key === "ArrowDown") {
-		e.preventDefault();
-		selectedIndex.value = (selectedIndex.value + 1) % filteredResults.value.length;
-		// Skip headers
-		if (filteredResults.value[selectedIndex.value]?.type === "header") {
-			onSearchKeydown(e);
-		}
-	} else if (e.key === "ArrowUp") {
-		e.preventDefault();
-		selectedIndex.value =
-			(selectedIndex.value - 1 + filteredResults.value.length) % filteredResults.value.length;
-		// Skip headers
-		if (filteredResults.value[selectedIndex.value]?.type === "header") {
-			onSearchKeydown(e);
-		}
-	} else if (e.key === "Enter" && selectedIndex.value !== -1) {
-		e.preventDefault();
-		selectItem(filteredResults.value[selectedIndex.value]);
-	}
-}
-
-function onSearchBlur() {
-	// Delay to allow click on results
-	setTimeout(() => {
-		showResults.value = false;
-		selectedIndex.value = -1;
-	}, 200);
+function onPaste() {
+	// ActionSelectorNode currently doesn't directly support paste triggering within the card,
+	// but we could implement it by emitting insert-node if needed.
+	// For now, we follow the existing pattern where paste is mainly for edge insertion.
 }
 
 function onCreate() {
@@ -245,7 +56,6 @@ function onCreate() {
 		selectedPreset.value.selected_label ||
 		action_type;
 	const nodeType = mapActionTypeToNodeType(action_type);
-	// Map Action Type to VueFlow node type using the same logic as App.vue
 	const node = store.nodes[nodeIndex];
 	if (!node) return;
 
@@ -298,14 +108,12 @@ function onCreate() {
 	}
 
 	// If the chosen node is terminal, remove any outgoing edges that might have existed
-	// (Selector nodes often have an outgoing edge if inserted on a connection)
 	if (isTerminalAction(action_type)) {
 		store.edges = store.edges.filter((edge) => edge.source !== props.id);
 		store.nodes[nodeIndex].data.next_step_if_true = null;
 		store.nodes[nodeIndex].data.next_step_if_false = null;
 	}
 
-	// Keep the user in the canvas flow: select the new node but do not auto-open config modal.
 	store.select(props.id);
 	store.show_sidebar = true;
 	store.touch_node(props.id);
@@ -315,10 +123,6 @@ function onCreate() {
 function deleteNode() {
 	frappe.confirm(__("Delete this node?"), () => store.delete_node(props.id));
 }
-
-onMounted(() => {
-	loadProcessOperations();
-});
 </script>
 
 <template>
@@ -343,18 +147,13 @@ onMounted(() => {
 			<div class="form-group search-group">
 				<label class="small text-muted">{{ __("Action Type") }}</label>
 				<div class="search-wrapper">
-					<div class="search-input-group">
-						<i class="fa fa-search search-icon"></i>
-						<input
-							type="text"
-							v-model="searchQuery"
-							class="form-control input-xs"
-							:placeholder="__('Search actions...')"
-							@focus="onSearchFocus"
-							@blur="onSearchBlur"
-							@keydown="onSearchKeydown"
-						/>
-					</div>
+					<ActionZone
+						:placeholder="__('Search actions...')"
+						:show-paste="false"
+						:auto-focus="false"
+						@select="onActionSelect"
+						@paste="onPaste"
+					/>
 					<div v-if="selectedPreset.operation" class="selected-operation-preview">
 						<span class="badge-chip">{{ selectedPreset.action_type }}</span>
 						<span class="selected-operation-text">
@@ -363,39 +162,6 @@ onMounted(() => {
 						<span v-if="selectedPreset.process_name" class="selected-process-name">
 							({{ selectedPreset.process_name }})
 						</span>
-					</div>
-					<div
-						v-if="showResults && filteredResults.length"
-						class="search-results"
-						@wheel.stop
-					>
-						<div
-							v-for="(item, idx) in filteredResults"
-							:key="idx"
-							:class="[
-								'search-result-item',
-								item.type === 'header' ? 'result-header' : 'result-option',
-								{ active: idx === selectedIndex },
-							]"
-							@mousedown.prevent="selectItem(item)"
-							@mouseover="selectedIndex = idx"
-						>
-							<template v-if="item.type === 'header'">
-								<span class="header-label">{{ item.label }}</span>
-							</template>
-							<template v-else>
-								<i
-									:class="['fa', item.icon?.replace('fa ', '')]"
-									:style="{ color: item.color }"
-								></i>
-								<div class="result-text">
-									<span class="result-label">{{ item.label }}</span>
-									<span v-if="item.description" class="result-desc">{{
-										item.description
-									}}</span>
-								</div>
-							</template>
-						</div>
 					</div>
 				</div>
 			</div>
@@ -424,7 +190,7 @@ onMounted(() => {
 
 <style scoped>
 .action-selector-card {
-	width: 220px;
+	width: 260px;
 	background: #fff;
 	border: 2px dashed #d1d8dd;
 	border-radius: 8px;
@@ -475,6 +241,26 @@ onMounted(() => {
 
 .node-body {
 	padding: 12px;
+	display: flex;
+	flex-direction: column;
+	max-height: 500px;
+}
+
+.search-wrapper {
+	max-height: 300px;
+	display: flex;
+	flex-direction: column;
+	border: 1px solid #e2e8f0;
+	border-radius: 6px;
+	padding-bottom: 4px;
+}
+
+.search-wrapper :deep(.popover-search) {
+	border-bottom: none;
+}
+
+.search-wrapper :deep(.popover-body) {
+	max-height: 200px;
 }
 
 .form-group {
@@ -486,54 +272,11 @@ onMounted(() => {
 	margin-bottom: 2px;
 }
 
-.search-wrapper {
-	position: relative;
-}
-
-.search-input-group {
-	position: relative;
-	display: flex;
-	align-items: center;
-}
-
-.search-icon {
-	position: absolute;
-	left: 8px;
-	color: #adb5bd;
-	font-size: 10px;
-	z-index: 1;
-	pointer-events: none;
-}
-
-.search-input-group input {
-	padding-left: 24px !important;
-}
-
-.search-results {
-	position: absolute;
-	top: 100%;
-	left: 0;
-	right: 0;
-	background: #fff;
-	border: 1px solid #e2e8f0;
-	border-radius: 6px;
-	box-shadow: 0 8px 24px rgba(0, 0, 0, 0.12);
-	z-index: 100;
-	max-height: 240px;
-	overflow-y: auto;
-	overscroll-behavior: contain;
-	margin-top: 2px;
-}
-
-.search-results :deep(*) {
-	user-select: none;
-}
-
 .selected-operation-preview {
 	display: flex;
 	align-items: center;
 	gap: 6px;
-	margin-top: 6px;
+	margin: 6px 8px 0;
 	flex-wrap: wrap;
 }
 
@@ -556,72 +299,6 @@ onMounted(() => {
 .selected-process-name {
 	font-size: 10px;
 	color: #6b7280;
-}
-
-.search-result-item {
-	cursor: pointer;
-	padding: 6px 10px;
-	font-size: 11px;
-	transition: background 0.15s;
-}
-
-.result-header {
-	font-size: 10px;
-	font-weight: 700;
-	text-transform: uppercase;
-	color: var(--text-muted);
-	padding: 8px 12px 4px;
-	cursor: default;
-	letter-spacing: 0.8px;
-	border-top: 1px solid #f1f5f9;
-	background: #fcfcfc;
-}
-
-.result-header:first-child {
-	border-top: none;
-}
-
-.result-option {
-	display: flex;
-	align-items: center;
-	padding: 8px 12px;
-	gap: 10px;
-	transition: all 0.2s;
-}
-
-.result-option:hover,
-.result-option.active {
-	background: #f3f4f6;
-}
-
-.result-option i {
-	font-size: 14px;
-	width: 16px;
-	text-align: center;
-}
-
-.result-text {
-	display: flex;
-	flex-direction: column;
-	min-width: 0;
-}
-
-.result-label {
-	font-weight: 500;
-	color: #111827;
-	white-space: nowrap;
-	overflow: hidden;
-	text-overflow: ellipsis;
-	line-height: 1.2;
-}
-
-.result-desc {
-	font-size: 10px;
-	color: #6b7280;
-	white-space: nowrap;
-	overflow: hidden;
-	text-overflow: ellipsis;
-	margin-top: 1px;
 }
 
 .node-footer {
@@ -669,33 +346,5 @@ onMounted(() => {
 .action-selector-card.selected .handle-target,
 .action-selector-card.selected .handle-source {
 	border-color: var(--primary) !important;
-}
-/* RTL Support */
-[dir="rtl"] .node-header {
-	flex-direction: row;
-}
-
-[dir="rtl"] .header-text {
-	text-align: right;
-}
-
-[dir="rtl"] .delete-btn {
-	margin-left: 0;
-	margin-right: auto;
-}
-
-[dir="rtl"] .search-icon {
-	left: auto;
-	right: 8px;
-}
-
-[dir="rtl"] .search-input-group input {
-	padding-left: 10px !important;
-	padding-right: 24px !important;
-}
-
-[dir="rtl"] .result-option i {
-	margin-right: 0;
-	margin-left: 10px;
 }
 </style>
